Organised code into namespaces
- PhunkyPhrogs.Core contains the PlayerController class,
- PhunkyPhrogs.Environment contains the PlatformSpawner and SectionMover class,
- PhunkyPhrogs.TrajectoryLine contains the TrajectoryManager and the DummyPlayerController.

PhunkyPhrogs is now the default namespace of the project.
