+++ Assets/Scripts/TrajectoryLine/TrajectoryManager.cs.meta
